+++ sdk/python/pulumi_postgresql/database.py
@@ -115,6 +115,10 @@ class Database(pulumi.CustomResource):
 
         __props__['template'] = template
 
+        if opts is None:
+            opts = pulumi.ResourceOptions()
+        if opts.version is None:
+            opts.version = utilities.get_version()
         super(Database, __self__).__init__(
             'postgresql:index/database:Database',
             resource_name,