+++ sdk/python/pulumi_postgresql/default_privileg.py
@@ -89,6 +89,10 @@ class DefaultPrivileg(pulumi.CustomResource):
             raise TypeError("Missing required property 'schema'")
         __props__['schema'] = schema
 
+        if opts is None:
+            opts = pulumi.ResourceOptions()
+        if opts.version is None:
+            opts.version = utilities.get_version()
         super(DefaultPrivileg, __self__).__init__(
             'postgresql:index/defaultPrivileg:DefaultPrivileg',
             resource_name,