+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 ## HEAD (Unreleased)
-___NULL___
+* Update to pulumi-terraform@013b95b1c
 
 ## 0.18.4 (2019-07-09)
 * Fix detailed diffs with nested computed values.